Clarifying warning on missing or unreadable file
If a new user doesn't have an `.env` file but instead relies on actual
environment settings on their platform (Heroku, Docker, etc) it
shouldn't appear that the user did something wrong. Clarifying the
warning is a good solution here.

If the file is unreadable, it's still theoretically possible the user is
relying on local environment variables, so that warning should also be
clarified.