Removed proxy variable feature. (#60)
Expanding variables automatically on a read is an anti-pattern.
Variable expansion by the shell should only be done when the
value is inserted inot the environment, but the vlaue should be
treated as opaque data.  Any processing or interpretation of the
variable should be done by the appliaction, not by the access
method.